% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.allTuplesSortedCpp}
\alias{.allTuplesSortedCpp}
\title{An internal helper function to check that each tuple is sorted in 
increasing order; only used for tuples of size > 2.}
\usage{
.allTuplesSortedCpp(pos1, internal_pos, posm)
}
\arguments{
\item{pos1}{An integer vector.}

\item{internal_pos}{An integer matrix.}

\item{posm}{An integer vector.}
}
\value{
TRUE if each tuple is sorted in strictly increasing order, 
FALSE otherwise.
}
\description{
An internal helper function to check that each tuple is sorted in 
increasing order; only used for tuples of size > 2.
}
\details{
.allTuplesSorted is adapted from http://stackoverflow.com/a/7601857.
Strict inequalities are required.
}
\keyword{internal}
