% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlaps-methods.R
\name{.findEqual.GTuples}
\alias{.findEqual.GTuples}
\title{An internal function used by the findOverlaps,GTuples,GTuples-method when 
type = "equal".}
\usage{
.findEqual.GTuples(query, subject, select, ignore.strand)
}
\arguments{
\item{query}{A GTuples instance}

\item{subject}{A GTuples instance}

\item{select}{See \code{findOverlaps} in the IRanges package for a 
description of this argument.}

\item{ignore.strand}{When set to \code{TRUE}, the strand information is
ignored in the overlap calculations.}
}
\description{
An internal function used by the findOverlaps,GTuples,GTuples-method when 
type = "equal".
}
\keyword{internal}
