% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtilities.R
\name{.GT2DT}
\alias{.GT2DT}
\title{Convert a GTuples object to a data.table.}
\usage{
.GT2DT(gt, ignore.strand = FALSE)
}
\arguments{
\item{gt}{A GTuples object}

\item{ignore.strand}{When set to \code{TRUE}, the strand is set to "*".}
}
\description{
Convert a GTuples object to a data.table.
}
\keyword{internal}
