% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.IPDCpp}
\alias{.IPDCpp}
\title{An internal function to compute the IPD of a GTuples; ; only used for 
tuples of size > 2.}
\usage{
.IPDCpp(pos1, internal_pos, posm)
}
\arguments{
\item{pos1}{An integer vector.}

\item{internal_pos}{An integer matrix.}

\item{posm}{An integer vector.}
}
\value{
An integer matrix with the same number of rows as internal_pos and 
number of columns equal to size - 1.
}
\description{
An internal function to compute the IPD of a GTuples; ; only used for 
tuples of size > 2.
}
\note{
This function silently coerces numeric matrices to integer matrices
and does integer subtraction. \emph{This will give unexpected results but
it's not a problem for me since I only use it on integer matrices.}
}
\keyword{internal}
