% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllUtilities.R
\name{.matrixDiffWithRecycling}
\alias{.matrixDiffWithRecycling}
\title{Compute column-wise difference of matrices with possibly different number of 
rows. Do this by iterating over columns, treating them as vectors and 
then using R's native vector recycling.}
\usage{
.matrixDiffWithRecycling(x, y)
}
\description{
Compute column-wise difference of matrices with possibly different number of 
rows. Do this by iterating over columns, treating them as vectors and 
then using R's native vector recycling.
}
\keyword{internal}
