% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.pcompareGTuplesCpp}
\alias{.pcompareGTuplesCpp}
\title{An internal function used to pcompare GTuples.}
\usage{
.pcompareGTuplesCpp(int_seqnames, int_strand, int_pos)
}
\arguments{
\item{int_seqnames}{An integer vector of length n. An integer 
representation of the difference in seqnames of each tuple.}

\item{int_strand}{An integer vector of length n. An integer representation 
of the difference in strand of each tuple.}

\item{int_pos}{An integer matrix with n rows. Each row represents the 
difference in positions of each tuple.}
}
\value{
An integer vector where each element is the comparison of a pair
of tuples. If the first tuple in the pair is "<" than the second tuple then 
the return value for that element is < 0, if the first tuple in the pair is 
"==" the second tuple then the return value is 0, and if the first tuple is 
">" that the second tuple then the return value is > 0.
}
\description{
An internal function used to pcompare GTuples.
}
\details{
The tuples should have already been converted to integer representations,
namely an integer vector for the difference in chromosome, an integer 
vector for the difference in strand and an integer matrix for the 
difference in  positions.
}
\keyword{internal}
