% Generated by roxygen2 (4.0.1): do not edit by hand
\name{IPD}
\alias{IPD}
\title{Compute the IPD of genomic tuples stored in a \code{\link{GTuples}} or
\code{\link{GTuplesList}} object.}
\usage{
IPD(x)
}
\arguments{
\item{x}{A \code{\link{GTuples}} object.}
}
\value{
If \code{size(x)} > 1, an integer matrix with the same number of
rows as \code{length(x)} and number of columns equal to \code{size(x) - 1}.
If \code{size(x)} == 1 or \code{size(x)} is \code{NA} then IPD is not
defined and an error is returned.
}
\description{
Compute the IPD of genomic tuples stored in a \code{\link{GTuples}} or
\code{\link{GTuplesList}} object.
}

