% Generated by roxygen2 (4.0.1): do not edit by hand
\name{size}
\alias{size}
\title{Get the size of genomic tuples stored in a \code{GTuples} object.}
\usage{
size(x)
}
\arguments{
\item{x}{A \code{\link{GTuples}} or \code{\link{GTuplesList}} object.}
}
\value{
An integer.
}
\description{
Get the size of genomic tuples stored in a \code{GTuples} object.
}

