% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{GenomicTuples-package}
\alias{GenomicTuples-package}
\title{Representation and manipulation of genomic tuples.}
\description{
\pkg{GenomicTuples} defines general purpose containers for storing genomic
tuples. It aims to provide functionality for tuples of genomic co-ordinates
that are analogous to those available for genomic ranges in the
GenomicRanges Bioconductor package.
}
\details{
Please refer to the vignettes to see how to use the  \pkg{GenomicTuples}
package.
}

