% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTuples.R
\docType{package}
\name{GenomicTuples-package}
\alias{GenomicTuples-package}
\title{Representation and manipulation of genomic tuples.}
\description{
\pkg{GenomicTuples} defines general purpose containers for storing genomic
tuples. It aims to provide functionality for tuples of genomic co-ordinates
that are analogous to those available for genomic ranges in the
GenomicRanges Bioconductor package.
}
\details{
Please refer to the vignettes to see how to use the \pkg{GenomicTuples}
package.
}
\references{
Peter F Hickey (2016). 
Representation and Manipulation of Genomic Tuples in R. 
\emph{JOSS}. URL \url{http://dx.doi.org/10.21105/joss.00020}
}

